Hoop_Net changes in Taher+Environment scene
--Ball Changes--
Added a Sphere Collider to the ball.
Turned off Mesh Collider to the ball.

--Hoop_Net Changes--
Made so the net uses Cloth when getting hit by the ball.
Added values to simulate wind to the net (if people don't like the wind so can I remove it later).

--Other--
Lost to the bot by 02-27 when testing values xD