Added new Winscreens, Load menu after win [not final?], Fade to Black between scenes
The automatic return to menu after a won game is just so that the game doesn't get "soft-locked-ish" after a victory (the game keeps going forever unless you restart the match or leave), I added a few fade to blacks when loading into menus,
currently no fade to black when loading into a match because the duration is currently too long which means the player misses out on match start. Perhaps we can create a shorter fade at some point? Anyway, if you're unhappy with the return to menu,
any sort of implementation of either winscreen or fade to black, or want me to change / remove / add something to it, just poke and prod at me and I'll get to it :)